Consolidate rollup configs
