+++ packages/algorithm/rollup.config.js
@@ -1,40 +1,6 @@
-import nodeResolve from '@rollup/plugin-node-resolve';
-import sourcemaps from 'rollup-plugin-sourcemaps';
-import postcss from 'rollup-plugin-postcss';
+import { createRollupConfig } from '../../rollup.src.config';
 
 const pkg = require('./package.json');
+const config = createRollupConfig(pkg);
 
-const globals = id =>
-  id.indexOf('@lumino/') === 0 ? id.replace('@lumino/', 'lumino_') : id;
-
-export default [
-  {
-    input: 'lib/index',
-    external: id => pkg.dependencies && !!pkg.dependencies[id],
-    output: [
-      {
-        file: pkg.main,
-        globals,
-        format: 'umd',
-        sourcemap: true,
-        name: pkg.name
-      },
-      {
-        file: pkg.module + '.js',
-        format: 'es',
-        sourcemap: true,
-        name: pkg.name
-      }
-    ],
-    plugins: [
-      nodeResolve({
-        preferBuiltins: true
-      }),
-      sourcemaps(),
-      postcss({
-        extensions: ['.css'],
-        minimize: true
-      })
-    ]
-  }
-];
+export default config;
