+++ packages/algorithm/tests/rollup.config.js
@@ -1,10 +1,1 @@
-import nodeResolve from '@rollup/plugin-node-resolve';
-import commonjs from '@rollup/plugin-commonjs';
-
-export default {
-  input: './lib/index.spec.js',
-  output: {
-    file: './lib/bundle.test.js'
-  },
-  plugins: [commonjs(), nodeResolve()]
-};
+export { default } from '../../../rollup.tests.config';
