+++ rollup.src.config.js
@@ -1,0 +1,38 @@
+import nodeResolve from '@rollup/plugin-node-resolve';
+import sourcemaps from 'rollup-plugin-sourcemaps';
+import postcss from 'rollup-plugin-postcss';
+
+const globals = id =>
+  id.indexOf('@lumino/') === 0 ? id.replace('@lumino/', 'lumino_') : id;
+
+export function createRollupConfig(pkg) {
+  return {
+    input: 'lib/index',
+    external: id => pkg.dependencies && !!pkg.dependencies[id],
+    output: [
+      {
+        file: pkg.main,
+        globals,
+        format: 'umd',
+        sourcemap: true,
+        name: pkg.name
+      },
+      {
+        file: pkg.module + '.js',
+        format: 'es',
+        sourcemap: true,
+        name: pkg.name
+      }
+    ],
+    plugins: [
+      nodeResolve({
+        preferBuiltins: true
+      }),
+      sourcemaps(),
+      postcss({
+        extensions: ['.css'],
+        minimize: true
+      })
+    ]
+  };
+}
